requirements.txt
orjson==3.11.2
pyahocorasick==2.2.0
